Fix the "one-page audit-log error"
Audit-logs are retrieved by REST-APIs. These are paged.
Until now, the Anonymizer has retrieved only one page (the first one) and has
searched for the anonymized user data only in this one page.
But in case of a lot of admin-activities, the anonymizing-data could be spread
over more than one page. If this is the case, retrieving the anonymized data
is incomplete.

With this fix, the Anonymizer iterates over all pages got between the begin and the end of one anonymization.