Get the user-name and user-key from the anonymization summary
Until Jira version 8.20 the event-API contained the "actionI18nKey" = "jira.auditing.user.renamed" where the anonymized user-name was got from.
But starting with Jira 8.21 this event is no longer part of an anonymization.

In older versions of the Anonymizer, it got the anonymized user-name from the
anonymization summary. This commit is to go back to that.